- clear files - upload flask app
