--- .gitignore
@@ -1,9 +1,0 @@
-__pycache__
-
-dataset/
-scripts/
-
-config/indoor_lidar_rgbd
-config/redwood_objects
-config/redwood_simulated
-config/stanford
