--- debug/pairwise_pc_alignment.py
@@ -1,55 +1,0 @@
-# ----------------------------------------------------------------------------
-# -                        Open3D: www.open3d.org                            -
-# ----------------------------------------------------------------------------
-# The MIT License (MIT)
-#
-# Copyright (c) 2018-2021 www.open3d.org
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
-# ----------------------------------------------------------------------------
-
-# examples/python/reconstruction_system/debug/pairwise_pc_alignment.py
-
-import argparse
-import json
-import sys
-
-pyexample_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.append(pyexample_path)
-from open3d_example import *
-
-sys.path.append(".")
-from initialize_config import *
-from register_fragments import *
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="mathching two point clouds")
-    parser.add_argument("config", help="reading json file for initial pose")
-    parser.add_argument("source_id", type=int, help="ID of source point cloud")
-    parser.add_argument("target_id", type=int, help="ID of target point cloud")
-    args = parser.parse_args()
-
-    with open(args.config) as json_file:
-        config = json.load(json_file)
-        initialize_config(config)
-        config['debug_mode'] = True
-        ply_file_names = get_file_list(
-            join(config["path_dataset"], config["folder_fragment"]), ".ply")
-        register_point_cloud_pair(ply_file_names, args.source_id,
-                                  args.target_id, config)
