--- sensors/azure_kinect_recorder.py
@@ -1,146 +1,0 @@
-# ----------------------------------------------------------------------------
-# -                        Open3D: www.open3d.org                            -
-# ----------------------------------------------------------------------------
-# The MIT License (MIT)
-#
-# Copyright (c) 2018-2021 www.open3d.org
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
-# ----------------------------------------------------------------------------
-
-# examples/python/reconstruction_system/sensors/azure_kinect_recorder.py
-
-import argparse
-import datetime
-import open3d as o3d
-
-
-class RecorderWithCallback:
-
-    def __init__(self, config, device, filename, align_depth_to_color):
-        # Global flags
-        self.flag_exit = False
-        self.flag_record = False
-        self.filename = filename
-
-        self.align_depth_to_color = align_depth_to_color
-        self.recorder = o3d.io.AzureKinectRecorder(config, device)
-        if not self.recorder.init_sensor():
-            raise RuntimeError('Failed to connect to sensor')
-
-    def escape_callback(self, vis):
-        self.flag_exit = True
-        if self.recorder.is_record_created():
-            print('Recording finished.')
-        else:
-            print('Nothing has been recorded.')
-        return False
-
-    def space_callback(self, vis):
-        if self.flag_record:
-            print('Recording paused. '
-                  'Press [Space] to continue. '
-                  'Press [ESC] to save and exit.')
-            self.flag_record = False
-
-        elif not self.recorder.is_record_created():
-            if self.recorder.open_record(self.filename):
-                print('Recording started. '
-                      'Press [SPACE] to pause. '
-                      'Press [ESC] to save and exit.')
-                self.flag_record = True
-
-        else:
-            print('Recording resumed, video may be discontinuous. '
-                  'Press [SPACE] to pause. '
-                  'Press [ESC] to save and exit.')
-            self.flag_record = True
-
-        return False
-
-    def run(self):
-        glfw_key_escape = 256
-        glfw_key_space = 32
-        vis = o3d.visualization.VisualizerWithKeyCallback()
-        vis.register_key_callback(glfw_key_escape, self.escape_callback)
-        vis.register_key_callback(glfw_key_space, self.space_callback)
-
-        vis.create_window('recorder', 1920, 540)
-        print("Recorder initialized. Press [SPACE] to start. "
-              "Press [ESC] to save and exit.")
-
-        vis_geometry_added = False
-        while not self.flag_exit:
-            rgbd = self.recorder.record_frame(self.flag_record,
-                                              self.align_depth_to_color)
-            if rgbd is None:
-                continue
-
-            if not vis_geometry_added:
-                vis.add_geometry(rgbd)
-                vis_geometry_added = True
-
-            vis.update_geometry(rgbd)
-            vis.poll_events()
-            vis.update_renderer()
-
-        self.recorder.close_record()
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Azure kinect mkv recorder.')
-    parser.add_argument('--config', type=str, help='input json kinect config')
-    parser.add_argument('--output', type=str, help='output mkv filename')
-    parser.add_argument('--list',
-                        action='store_true',
-                        help='list available azure kinect sensors')
-    parser.add_argument('--device',
-                        type=int,
-                        default=0,
-                        help='input kinect device id')
-    parser.add_argument('-a',
-                        '--align_depth_to_color',
-                        action='store_true',
-                        help='enable align depth image to color')
-    args = parser.parse_args()
-
-    if args.list:
-        o3d.io.AzureKinectSensor.list_devices()
-        exit()
-
-    if args.config is not None:
-        config = o3d.io.read_azure_kinect_sensor_config(args.config)
-    else:
-        config = o3d.io.AzureKinectSensorConfig()
-
-    if args.output is not None:
-        filename = args.output
-    else:
-        filename = '{date:%Y-%m-%d-%H-%M-%S}.mkv'.format(
-            date=datetime.datetime.now())
-    print('Prepare writing to {}'.format(filename))
-
-    device = args.device
-    if device < 0 or device > 255:
-        print('Unsupported device id, fall back to 0')
-        device = 0
-
-    r = RecorderWithCallback(config, device, filename,
-                             args.align_depth_to_color)
-    r.run()
