+++ static/css/style.css
@@ -1,0 +1,43 @@
+body{
+    text-align: center;
+}
+
+h1{
+    text-decoration: underline;
+    text-decoration-style: wavy;
+    text-decoration-color: red;
+    font-family: fantasy;
+    background-color:aquamarine
+}
+
+#pipeline, #preview{
+    text-align: center;
+    margin-top: 1px;
+    margin-left: auto;
+    margin-right:auto;
+    width:75%;
+    height:50%;
+}
+
+
+#pipeline{
+    background-color: aqua;
+    padding: 30px 30px;
+}
+
+#pipeline button{
+    display: block;
+    margin-left: auto;
+    margin-right:auto;
+    margin-top: 1em;
+    
+}
+
+#preview{
+    background-color: aquamarine;
+    padding: 30px 30px;
+}
+
+.navbar{
+    background-color:rgb(171, 243, 219)
+}
